requirements.txt
annotated-types==0.7.0
anyio==4.11.0
async-timeout==4.0.3
certifi==2025.10.5
charset-normalizer==3.4.4
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.79
langchain-text-splitters==0.3.11
langsmith==0.4.37
mypy_extensions==1.1.0
orjson==3.11.4
packaging==25.0
pydantic==2.12.3
pydantic_core==2.41.4
PyYAML==6.0.3
qwikidata==0.4.2
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.44
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
Wikipedia-API==0.8.1
zstandard==0.25.0
